Add hook to skip shutdown hook in command tests
This commit adds a hook for skipping the shutdown hook in command tests,
otherwise we install a bunch of unneeded hooks during shutdown tests,
and need special permissions to do so anyway.
