+++ test/framework/src/main/java/org/elasticsearch/bootstrap/ESElasticsearchCliTestCase.java
@@ -50,6 +50,11 @@ abstract class ESElasticsearchCliTestCase extends ESTestCase {
                     init.set(true);
                     initConsumer.accept(!daemonize, pidFile, quiet, esSettings);
                 }
+
+                @Override
+                protected boolean addShutdownHook() {
+                    return false;
+                }
             }, terminal);
             assertThat(status, equalTo(expectedStatus));
             assertThat(init.get(), equalTo(expectedInit));
